Extract the amount converter from the Gateway
