+++ app/models/solidus_stripe/gateway.rb
@@ -7,6 +7,8 @@ require 'stripe'
 # - https://stripe.com/docs/charges/placing-a-hold
 module SolidusStripe
   class Gateway
+    include SolidusStripe::MoneyToStripeAmountConverter
+
     def initialize(options)
       # Cannot use kwargs because of how the Gateway is initialized by Solidus.
       @client = Stripe::StripeClient.new(
@@ -76,7 +78,7 @@ module SolidusStripe
       # Charge the Customer instead of the card:
       payment_intent, _response = client.request do
         Stripe::PaymentIntent.create({
-          amount: MoneyToStripeAmountConverter.to_stripe_amount(amount_in_cents, currency),
+          amount: to_stripe_amount(amount_in_cents, currency),
           currency: currency,
           **options[:payment_intent_options].to_h
         })
@@ -128,7 +130,7 @@ module SolidusStripe
 
       refund, _response = client.request do
         Stripe::Refund.create(
-          amount: MoneyToStripeAmountConverter.to_stripe_amount(amount_in_cents, currency),
+          amount: to_stripe_amount(amount_in_cents, currency),
           payment_intent: source.stripe_payment_intent_id,
         )
       end
@@ -136,88 +138,6 @@ module SolidusStripe
       ActiveMerchant::Billing::Response.new(
         true, "PaymentIntent was refunded successfully", { 'stripe_refund' => refund.to_json }, {}
       )
-    end
-
-    module MoneyToStripeAmountConverter
-      extend self
-
-      ZERO_DECIMAL_CURRENCIES = %w[
-        BIF
-        CLP
-        DJF
-        GNF
-        JPY
-        KMF
-        KRW
-        MGA
-        PYG
-        RWF
-        UGX
-        VND
-        VUV
-        XAF
-        XOF
-        XPF
-      ].freeze
-
-      THREE_DECIMAL_CURRENCIES = %w[
-        BHD
-        JOD
-        KWD
-        OMR
-        TND
-      ].freeze
-
-      # special currencies that are represented in cents but
-      # should be divisible by 100, thus making them integer only.
-      DIVISIBLE_BY_100 = %w[
-        HUF
-        TWD
-        UGX
-      ].freeze
-
-      # Solidus will provide a "fractional" amount, that is specific for each currency
-      # following the configurationo defined in the Money gem.
-      #
-      # Stripe uses the "smallest currency unit",
-      # (e.g., 100 cents to charge $1.00 or 100 to charge ¥100, a zero-decimal currency)
-      # https://stripe.com/docs/currencies#zero-decimal
-      #
-      # We need to ensure the fractional amount is considering the same number of decimals.
-      def to_stripe_amount(fractional, currency)
-        solidus_subunit_to_unit, stripe_subunit_to_unit = subunit_to_unit(currency)
-
-        if stripe_subunit_to_unit == solidus_subunit_to_unit
-          fractional
-        else
-          (fractional / solidus_subunit_to_unit.to_d) * stripe_subunit_to_unit.to_d
-        end
-      end
-
-      def to_solidus_amount(fractional, currency)
-        solidus_subunit_to_unit, stripe_subunit_to_unit = subunit_to_unit(currency)
-
-        if stripe_subunit_to_unit == solidus_subunit_to_unit
-          fractional
-        else
-          (fractional / stripe_subunit_to_unit.to_d) * solidus_subunit_to_unit.to_d
-        end
-      end
-
-      private
-
-      def subunit_to_unit(currency)
-        solidus_subunit_to_unit = ::Money::Currency.new(currency).subunit_to_unit
-        stripe_subunit_to_unit =
-          case currency.to_s.upcase
-          when *ZERO_DECIMAL_CURRENCIES then 1
-          when *THREE_DECIMAL_CURRENCIES then 1000
-          when *DIVISIBLE_BY_100 then 100
-          else 100
-          end
-
-        return solidus_subunit_to_unit, stripe_subunit_to_unit
-      end
     end
   end
 end
